+++ src/components/tab.d.ts
@@ -1,5 +1,5 @@
 import {Entity} from "./entity";
-import App from "./app";
+import App from "../app";
 
 export default interface TabComponent {
 	/**